feat: allow usage of local project files

+++ src/actions.rs
@@ -43,19 +43,22 @@ pub enum Error {
     UnsupportedFileExtension { extension: OsString },
 }
 
-pub fn start_project<S: AsRef<OsStr>>(
+pub fn start_project(
     config: &Config,
-    project_name: S,
-    template: Option<&str>,
+    project_name: Option<OsString>,
+    template_file: Option<OsString>,
     force_attach: Option<bool>,
     show_source: bool,
     verbose: bool,
     args: Vec<String>,
 ) -> Result<(), Box<dyn error::Error>> {
-    let project = project::load(config, &project_name, force_attach, &args)?;
+    let (project_name, project_file) = project::get_filename(config, project_name)?;
+    ensure!(project_file.is_file(), ProjectDoesNotExist { project_name });
+
+    let project = project::load(config, &project_name, &project_file, force_attach, &args)?;
     project.check()?;
 
-    // Build and run tmux commands
+    // Build the tmux source file from a template file
     let mut context = Context::new();
     context.insert("project", &project);
     context.insert("verbose", &verbose);
@@ -63,35 +66,14 @@ pub fn start_project<S: AsRef<OsStr>>(
     let template_comand = project.get_tmux_command_for_template()?;
     context.insert("tmux_command", &template_comand);
 
-    let mut tera = Tera::default();
-    tera.register_filter("quote", source::QuoteFilter {});
-
-    let template_content: String;
-    let template = match &project.template {
-        ProjectTemplate::Raw(content) => {
-            template_content = content.into();
-            template_content.as_str()
-        }
-        ProjectTemplate::File(filename) => {
-            let full_path = if filename.has_root() {
-                filename.to_owned()
-            } else {
-                let project_dir = config
-                    .get_projects_dir(project_name.as_ref())?
-                    .parent()
-                    .map_or_else(|| PathBuf::new(), |p| PathBuf::from(p));
-                PathBuf::from(project_dir.join(filename).canonicalize()?)
-            };
-
-            template_content = fs::read_to_string(full_path)?;
-            template_content.as_str()
-        }
-        ProjectTemplate::Default => {
-            template.unwrap_or(include_str!("assets/default_template.tera"))
-        }
+    let template_content = match template_file {
+        Some(template_file) => fs::read_to_string(template_file)?,
+        None => project::get_template(&project, project_file)?,
     };
 
-    let source = tera.render_str(template, &context)?;
+    let mut tera = Tera::default();
+    tera.register_filter("quote", source::QuoteFilter {});
+    let source = tera.render_str(template_content.as_str(), &context)?;
 
     // Run tmux
     if show_source {
@@ -160,12 +142,15 @@ pub fn start_project<S: AsRef<OsStr>>(
     Ok(())
 }
 
-pub fn kill_project<S: AsRef<OsStr>>(
+pub fn kill_project(
     config: &Config,
-    project_name: S,
+    project_name: Option<OsString>,
     args: Vec<String>,
 ) -> Result<(), Box<dyn error::Error>> {
-    let project = project::load(config, &project_name, None, &args)?;
+    let (project_name, project_file) = project::get_filename(config, project_name)?;
+    ensure!(project_file.is_file(), ProjectDoesNotExist { project_name });
+
+    let project = project::load(config, &project_name, &project_file, None, &args)?;
     project.check()?;
 
     let session_name = project
@@ -177,7 +162,7 @@ pub fn kill_project<S: AsRef<OsStr>>(
     let (tmux_command, tmux_args) = project.get_tmux_command(vec![
         OsString::from("kill-session"),
         OsString::from("-t"),
-        OsString::from(session_name.to_owned()),
+        OsString::from(&session_name),
     ])?;
 
     let status = Command::new(tmux_command).args(tmux_args).spawn()?.wait()?;
@@ -192,68 +177,65 @@ pub fn kill_project<S: AsRef<OsStr>>(
     Ok(())
 }
 
-pub fn edit_project<S1: AsRef<OsStr>, S2: AsRef<OsStr>, S3: AsRef<OsStr>>(
+pub fn edit_project<S>(
     config: &Config,
-    project_name: S1,
-    extension: S2,
-    editor: S3,
+    project_name: Option<OsString>,
+    extension: Option<OsString>,
+    editor: S,
     no_check: bool,
     args: Vec<String>,
-) -> Result<(), Box<dyn error::Error>> {
-    let project_name = project_name.as_ref();
-    ensure!(!project_name.is_empty(), ProjectNameEmpty {});
+) -> Result<(), Box<dyn error::Error>>
+where
+    S: AsRef<OsStr>,
+{
+    let (project_name, project_file) = project::get_filename(config, project_name)?;
 
+    let extension = extension.unwrap_or(OsString::from("yml"));
     edit::check_supported_extension(&extension)?;
-    let project_name = edit::add_extension_to_project_name(&project_name, &extension);
+    let project_file = project_file.with_extension(extension);
 
     let editor = editor.as_ref();
     ensure!(!editor.is_empty(), EditorEmpty {});
 
     // Make sure the project's parent directory exists
-    let namespace = utils::get_project_namespace(&project_name)?;
-    let data_dir = config.get_projects_dir("")?;
-    mkdirp(data_dir.join(&namespace))?;
+    if let Some(parent) = project_file.parent() {
+        mkdirp(parent)?;
+    }
 
     // Make sure the project's yml file exists
-    let project_path = data_dir.join(&project_name);
-    let project_path = project::test_for_file_extensions(project_path)?;
-
     ensure!(
-        !project_path.is_dir(),
-        ProjectFileIsADirectory { path: project_path }
+        !project_file.is_dir(),
+        ProjectFileIsADirectory { path: project_file }
     );
-    if !project_path.exists() {
-        edit::create_project(&project_name, &project_path)?;
+
+    if !project_file.exists() {
+        edit::create_project(&project_name, &project_file)?;
     }
 
     // Open it with editor
-    let (command, command_args) = utils::parse_command(editor, &[OsString::from(project_path)])?;
+    let editor = OsString::from(editor);
+    let (command, command_args) =
+        utils::parse_command(editor.as_os_str(), &[OsString::from(&project_file)])?;
     let mut child = Command::new(command).args(command_args).spawn()?;
 
     if !no_check {
         child.wait()?;
 
         // Perform a check on the project
-        let project = project::load(config, project_name, None, &args)?;
+        let project = project::load(config, &project_name, &project_file, None, &args)?;
         project.check()?;
     }
 
     Ok(())
 }
 
-pub fn remove_project<S: AsRef<OsStr>>(
+pub fn remove_project(
     config: &Config,
-    project_name: S,
+    project_name: Option<OsString>,
     no_input: bool,
 ) -> Result<(), Box<dyn error::Error>> {
-    let project_name = project_name.as_ref();
-    ensure!(!project_name.is_empty(), ProjectNameEmpty {});
-
-    let projects_dir = config.get_projects_dir("")?;
-    let project_path = projects_dir.join(project_name);
-    let project_path = project::test_for_file_extensions(project_path)?;
-
-    ensure!(project_path.is_file(), ProjectDoesNotExist { project_name });
+    let (project_name, project_file) = project::get_filename(config, project_name)?;
+    ensure!(project_file.is_file(), ProjectDoesNotExist { project_name });
 
     if !no_input
         && !Confirm::new()
@@ -269,13 +251,18 @@ pub fn remove_project<S: AsRef<OsStr>>(
         return Ok(());
     }
 
-    fs::remove_file(&project_path)?;
-    for parent in project_path.ancestors() {
-        if parent == projects_dir {
-            break;
-        }
+    fs::remove_file(&project_file)?;
 
-        let _ = fs::remove_dir(parent);
+    // If it's in the projects directory, remove parent directories that are empty
+    let projects_dir = config.get_projects_dir("")?;
+    if project_file.starts_with(&projects_dir) {
+        for parent in project_file.ancestors() {
+            if parent == projects_dir {
+                break;
+            }
+
+            let _ = fs::remove_dir(parent);
+        }
     }
 
     println!("Project {:?} removed successfully.", project_name);
@@ -301,21 +288,79 @@ pub fn list_projects(config: &Config) -> Result<(), Box<dyn error::Error>> {
 mod project {
     use super::*;
 
-    pub fn load<S: AsRef<OsStr>>(
+    pub fn get_filename<S>(
+        config: &Config,
+        project_filename: Option<S>,
+    ) -> Result<(OsString, PathBuf), Box<dyn error::Error>>
+    where
+        S: AsRef<OsStr>,
+    {
+        match project_filename {
+            Some(project_name) => {
+                let project_name = project_name.as_ref();
+                ensure!(!project_name.is_empty(), ProjectNameEmpty {});
+
+                let projects_dir = config.get_projects_dir("")?;
+                let project_file = projects_dir.join(project_name);
+                let project_file = test_for_file_extensions(project_file)?;
+
+                Ok((OsString::from(project_name), project_file))
+            }
+            None => {
+                // Try to find a local project file in current directory and all ancestors
+                let mut project_dir = std::env::current_dir()?;
+                loop {
+                    let project_file = project_dir.join(PathBuf::from(".rmux"));
+
+                    // Try for each supported file extension
+                    for ext in FILE_EXTENSIONS {
+                        let project_file = project_file.with_extension(ext);
+                        if project_file.exists() && !project_file.is_dir() {
+                            let project_name = project_dir
+                                .file_name()
+                                .map_or(OsString::new(), |p| p.to_os_string());
+                            return Ok((project_name, project_file));
+                        }
+                    }
+
+                    // Move on to parent if nothing is found
+                    match project_dir.parent() {
+                        Some(parent_dir) => project_dir = parent_dir.to_path_buf(),
+                        None => break,
+                    }
+                }
+
+                // Fall back to local project file
+                let project_dir = std::env::current_dir()?;
+                let project_file = project_dir.join(".rmux.yml");
+                let project_name = match project_dir.file_name() {
+                    Some(name) => {
+                        // Remove dots and colons
+                        let unicode_name = name.to_string_lossy();
+                        OsString::from(unicode_name.replace(".", "").replace(":", ""))
+                    }
+                    None => OsString::new(),
+                };
+
+                Ok((project_name, project_file))
+            }
+        }
+    }
+
+    pub fn load<S, P>(
         config: &Config,
         project_name: S,
+        project_file: P,
         force_attach: Option<bool>,
         args: &[String],
-    ) -> Result<Project, Box<dyn error::Error>> {
+    ) -> Result<Project, Box<dyn error::Error>>
+    where
+        S: AsRef<OsStr>,
+        P: AsRef<Path>,
+    {
         let project_name = project_name.as_ref();
-        ensure!(!project_name.is_empty(), ProjectNameEmpty {});
 
-        let projects_dir = config.get_projects_dir("")?;
-        let project_path = projects_dir.join(project_name);
-        let project_path = test_for_file_extensions(project_path)?;
-        ensure!(project_path.is_file(), ProjectDoesNotExist { project_name });
-
-        let project_yaml = fs::read_to_string(project_path)?;
+        let project_yaml = fs::read_to_string(project_file)?;
         let project_yaml = env_with_context(&project_yaml, |s| env_context(s, args))
             .map_err(|x| x.to_string())?
             .to_string();
@@ -339,9 +384,10 @@ mod project {
         Ok(env::var(s).ok())
     }
 
-    pub fn test_for_file_extensions<P: AsRef<Path>>(
-        path: P,
-    ) -> Result<PathBuf, Box<dyn error::Error>> {
+    pub fn test_for_file_extensions<P>(path: P) -> Result<PathBuf, Box<dyn error::Error>>
+    where
+        P: AsRef<Path>,
+    {
         let path = path.as_ref();
 
         // If the path already contains an extension, use it directly
@@ -360,6 +406,33 @@ mod project {
 
         // If no file was found, fall back to the first extension in the list
         return Ok(path.with_extension(FILE_EXTENSIONS[0]));
+    }
+
+    pub fn get_template<P>(
+        project: &Project,
+        project_file: P,
+    ) -> Result<String, Box<dyn error::Error>>
+    where
+        P: AsRef<Path>,
+    {
+        let project_file = project_file.as_ref();
+
+        Ok(match &project.template {
+            ProjectTemplate::Raw(content) => content.to_owned(),
+            ProjectTemplate::File(filename) => {
+                let full_path = if filename.has_root() {
+                    filename.to_owned()
+                } else {
+                    let project_dir = project_file
+                        .parent()
+                        .map_or_else(|| PathBuf::new(), |p| PathBuf::from(p));
+                    PathBuf::from(project_dir.join(filename).canonicalize()?)
+                };
+
+                fs::read_to_string(full_path)?
+            }
+            ProjectTemplate::Default => include_str!("assets/default_template.tera").to_string(),
+        })
     }
 }
 
@@ -426,10 +499,14 @@ mod source {
 mod edit {
     use super::*;
 
-    pub fn create_project<S: AsRef<OsStr>, P: AsRef<Path>>(
+    pub fn create_project<S, P>(
         project_name: S,
         project_path: P,
-    ) -> Result<(), Box<dyn error::Error>> {
+    ) -> Result<(), Box<dyn error::Error>>
+    where
+        S: AsRef<OsStr>,
+        P: AsRef<Path>,
+    {
         let project_name = project_name.as_ref();
         let project_name = strip_extension_from_project_name(project_name);
         let default_project_yml = include_str!("assets/default_project.yml")
@@ -443,9 +520,10 @@ mod edit {
         Ok(())
     }
 
-    pub fn check_supported_extension<S: AsRef<OsStr>>(
-        extension: S,
-    ) -> Result<(), Box<dyn error::Error>> {
+    pub fn check_supported_extension<S>(extension: S) -> Result<(), Box<dyn error::Error>>
+    where
+        S: AsRef<OsStr>,
+    {
         let extension = extension.as_ref();
 
         ensure!(
@@ -459,22 +537,21 @@ mod edit {
     }
 
     // Disguise the project name as a Path for easy access to .with_extension()
-    pub fn strip_extension_from_project_name<P: AsRef<Path>>(project_name: P) -> OsString {
+    pub fn strip_extension_from_project_name<P>(project_name: P) -> OsString
+    where
+        P: AsRef<Path>,
+    {
         OsString::from(project_name.as_ref().with_extension(""))
-    }
-
-    pub fn add_extension_to_project_name<P: AsRef<Path>, S: AsRef<OsStr>>(
-        project_name: P,
-        extension: S,
-    ) -> OsString {
-        OsString::from(project_name.as_ref().with_extension(extension))
     }
 }
 
 mod list {
     use super::*;
 
-    pub fn get_projects<P: AsRef<Path>>(path: P) -> Result<Vec<OsString>, Box<dyn error::Error>> {
+    pub fn get_projects<P>(path: P) -> Result<Vec<OsString>, Box<dyn error::Error>>
+    where
+        P: AsRef<Path>,
+    {
         let path = path.as_ref();
         let mut projects = vec![];
